fixed units conversion

+++ lunabot_embedded/firmware/teensy_main/lib/lunabot_drivers/interfaces.hpp
@@ -107,7 +107,7 @@ class Encoder_Bus {
 public:
   Encoder_Bus() {};
   static void init(uint8_t is_top);
-  static float read(uint8_t id);
+  static long read(uint8_t id);
 
 private:
   static constexpr int NUM_ACTUATORS = 2;
